+++ svelte.config.js
@@ -8,7 +8,8 @@ const config = {
 	},
 	kit: {
 		adapter: adapter({
-			runtime: 'nodejs22.x'
+			runtime: 'nodejs22.x',
+			maxDuration: 300
 		})
 	}
 };